requirements.txt
pandas==1.5.3
google-api-python-client==1.7.2
google-auth==2.16.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1